Made custom format instead of rdf patch
