+++ timbuctoo-instancev4/src/test/java/nl/knaw/huygens/timbuctoo/v5/rdfio/implementations/rdf4j/Rdf4jNQuadUdParserTest.java
@@ -1,0 +1,142 @@
+package nl.knaw.huygens.timbuctoo.v5.rdfio.implementations.rdf4j;
+
+import nl.knaw.huygens.timbuctoo.v5.dataset.RdfProcessor;
+import nl.knaw.huygens.timbuctoo.v5.filestorage.dto.CachedLog;
+import org.junit.Test;
+
+import javax.ws.rs.core.MediaType;
+import java.io.File;
+import java.io.IOException;
+import java.io.Reader;
+import java.io.StringReader;
+import java.util.Optional;
+
+import static org.mockito.ArgumentMatchers.anyBoolean;
+import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.verify;
+
+public class Rdf4jNQuadUdParserTest {
+
+  private static final String START_FROM = "";
+  private static final String CURSOR_PREFIX = "";
+
+  @Test
+  public void itDeletes() throws Exception {
+    RdfProcessor rdfProcessor = mock(RdfProcessor.class);
+    StringReader reader =
+      new StringReader("-<http://example.org/subject1> <http://pred> \"12\"^^<http://number> <http://some_graph> .");
+    CachedLog cachedLog = rdfPatchLog(reader);
+    Rdf4jNQuadUdParser instance = new Rdf4jNQuadUdParser();
+
+    instance.importRdf(CURSOR_PREFIX, START_FROM, cachedLog, rdfProcessor);
+
+    verify(rdfProcessor).onQuad(
+      false,
+      "0",
+      "http://example.org/subject1",
+      "http://pred",
+      "12",
+      "http://number",
+      null,
+      "http://some_graph"
+    );
+  }
+
+  @Test
+  public void itAdds() throws Exception {
+    RdfProcessor rdfProcessor = mock(RdfProcessor.class);
+    StringReader reader =
+      new StringReader("+<http://example.org/subject1> <http://pred> \"12\"^^<http://number> <http://some_graph> .");
+    CachedLog cachedLog = rdfPatchLog(reader);
+    Rdf4jNQuadUdParser instance = new Rdf4jNQuadUdParser();
+
+    instance.importRdf(CURSOR_PREFIX, START_FROM, cachedLog, rdfProcessor);
+
+    verify(rdfProcessor).onQuad(
+      true,
+      "0",
+      "http://example.org/subject1",
+      "http://pred",
+      "12",
+      "http://number",
+      null,
+      "http://some_graph"
+    );
+  }
+
+  @Test
+  public void itIgnoresLinesThatStartWithoutAPlusOrAMinus() throws Exception {
+    RdfProcessor rdfProcessor = mock(RdfProcessor.class);
+    StringReader reader = new StringReader(
+      " <http://example.org/subject1> <http://pred> \"12\"^^<http://number> <http://some_graph> .\n" +
+        "@@ -1,4 +1,4 @@"
+    );
+    CachedLog cachedLog = rdfPatchLog(reader);
+    Rdf4jNQuadUdParser instance = new Rdf4jNQuadUdParser();
+
+    instance.importRdf(CURSOR_PREFIX, START_FROM, cachedLog, rdfProcessor);
+
+    verify(rdfProcessor, never()).onQuad(
+      anyBoolean(), anyString(), anyString(), anyString(), anyString(), anyString(), anyString(), anyString()
+    );
+  }
+
+  @Test
+  public void itIgnoresLinesThatStartWithTriplePlusSigns() throws Exception {
+    RdfProcessor rdfProcessor = mock(RdfProcessor.class);
+    StringReader reader = new StringReader("+++ fruits2\t2017-08-16 11:38:05.327645535 +0200");
+    CachedLog cachedLog = rdfPatchLog(reader);
+    Rdf4jNQuadUdParser instance = new Rdf4jNQuadUdParser();
+
+    instance.importRdf(CURSOR_PREFIX, START_FROM, cachedLog, rdfProcessor);
+
+    verify(rdfProcessor, never()).onQuad(
+      anyBoolean(), anyString(), anyString(), anyString(), anyString(), anyString(), anyString(), anyString()
+    );
+  }
+
+  @Test
+  public void itIgnoresLinesThatStartWithTripleMinusSigns() throws Exception {
+    RdfProcessor rdfProcessor = mock(RdfProcessor.class);
+    StringReader reader = new StringReader("--- fruits1\t2017-08-16 11:37:47.247741827 +0200");
+    CachedLog cachedLog = rdfPatchLog(reader);
+    Rdf4jNQuadUdParser instance = new Rdf4jNQuadUdParser();
+
+    instance.importRdf(CURSOR_PREFIX, START_FROM, cachedLog, rdfProcessor);
+
+    verify(rdfProcessor, never()).onQuad(
+      anyBoolean(), anyString(), anyString(), anyString(), anyString(), anyString(), anyString(), anyString()
+    );
+  }
+
+  private CachedLog rdfPatchLog(StringReader reader) {
+    return new CachedLog() {
+      @Override
+      public void close() throws Exception {
+
+      }
+
+      @Override
+      public String getName() {
+        return "http://example.com";
+      }
+
+      @Override
+      public File getFile() {
+        throw new UnsupportedOperationException("Method should not be needed");
+      }
+
+      @Override
+      public Reader getReader() throws IOException {
+        return reader;
+      }
+
+      @Override
+      public Optional<MediaType> getMimeType() {
+        return Optional.of(new MediaType("application", "rdf-patch"));
+      }
+    };
+  }
+}
